change src folder structure
