+++ src/components/FilterAndSearch.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
-import { media } from '../../theme/globalStyle'
+import { media } from '../theme/globalStyle'
 
 // Filter component ------------------------------------
 const StyledFilter = styled.div`
