+++ src/components/Header.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import Link from 'gatsby-link'
 import styled from 'styled-components'
 
-import { StyledH1, media } from '../../theme/globalStyle'
+import { StyledH1, media } from '../theme/globalStyle'
 import PageNav from './PageNav.js'
 
 const StyledHeader = styled.div`
