+++ src/components/JobsContainer.js
@@ -3,12 +3,13 @@ import styled from 'styled-components'
 import { withTheme } from 'styled-components'
 import PropTypes from 'prop-types'
 
-import { media } from '../../theme/globalStyle'
-import FilterAndSearch from './FilterAndSearch'
+import { media } from '../theme/globalStyle'
+import JobsFilter from './JobsFilter'
 import JobsList from './JobsList'
 import Pagination from './Pagination'
+import FilterAndSearch from './FilterAndSearch'
 import { LoadingContent } from './Titles'
-import { fakeLearnAPI, fakeLearnAPISearch } from '../utils/api'
+import { fakeAPI, fakeAPISearch } from '../utils/api'
 
 const Wrapper = styled.div`
   display: grid;
@@ -16,21 +17,19 @@ const Wrapper = styled.div`
   grid-template-rows: auto;
   grid-column-gap: 1rem;
   grid-template-areas:
-    '. . fs fs fs fs fs fs fs fs . .'
-    '. . lst lst lst lst lst lst lst lst . .';
+    '. fs fs fs fs fs fs fs fs fs fs .'
+    '. filt filt filt lst lst lst lst lst lst lst .';
   margin-top: 2.5rem;
-  ${media.giant`
+  ${media.tablet`
     grid-column-gap: 0.5rem;
     grid-template-areas:
-      '. fs fs fs fs fs fs fs fs fs fs .'
-      '. lst lst lst lst lst lst lst lst lst lst .';
-  `} ${media.tablet`
-    grid-template-areas:
       'fs fs fs fs fs fs fs fs fs fs fs fs'
-      '. lst lst lst lst lst lst lst lst lst lst .';
+      'filt filt filt filt lst lst lst lst lst lst lst .';
   `} ${media.phone`
+    grid-column-gap: 0.5rem;
     grid-template-areas:
       'fs fs fs fs fs fs fs fs fs fs fs fs'
+      'filt filt filt filt filt filt filt filt filt filt filt filt'
       '. lst lst lst lst lst lst lst lst lst lst .';
   `};
 `
@@ -44,23 +43,25 @@ const List = styled.div`
   padding: 0;
 `
 
-class CoursesContainer extends React.Component {
+class JobsContainer extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
       loading: true,
       searchQuery: '',
+      filter: ['any', 'any', 'any', 'any'],
       query: [
         {
           title: '',
-          date: '',
-          period: '',
-          level: '',
-          topic: '',
+          type: '',
+          role: '',
+          location: '',
           description: ''
         }
       ]
     }
+    this.handleChange = this.handleChange.bind(this)
+    this.submitQuery = this.submitQuery.bind(this)
     this.changePage = this.changePage.bind(this)
     this.handleDates = this.handleDates.bind(this)
     this.handleSearch = this.handleSearch.bind(this)
@@ -69,7 +70,21 @@ class CoursesContainer extends React.Component {
 
   componentDidMount() {
     const loading = false
-    fakeLearnAPI().then(query => this.setState({ loading, query }))
+    fakeAPI().then(query => this.setState({ loading, query }))
+  }
+
+  handleChange(num, data) {
+    let arr = this.state.filter.slice()
+    arr[num - 1] = data
+
+    this.setState({
+      filter: arr
+    })
+  }
+
+  submitQuery() {
+    const data = this.state.filter
+    fakeAPI(10, data).then(query => this.setState({ query }))
   }
 
   changePage(num) {
@@ -78,9 +93,9 @@ class CoursesContainer extends React.Component {
       var result
       if (this.state.searchQuery) {
         const searchQuery = this.state.searchQuery
-        result = fakeLearnAPI(undefined, searchQuery)
+        result = fakeAPI(undefined, undefined, searchQuery)
       } else {
-        result = fakeLearnAPI()
+        result = fakeAPI()
       }
       result.then(query => this.setState({ query }))
     }
@@ -99,16 +114,20 @@ class CoursesContainer extends React.Component {
 
   handleDates(str) {
     // immitation of new query
+    const data = this.state.filter
     const currLength =
         this.state.query.length < 5 ? 7 : this.state.query.length,
-      length = str === 'all' ? 7 : Math.floor(Math.random() * (currLength + 1))
+      length =
+        str === 'all'
+          ? 7
+          : Math.floor(Math.random() * (currLength + 1))
     var result
 
     if (this.state.searchQuery) {
       const searchQuery = this.state.searchQuery
-      result = fakeLearnAPI(length, searchQuery)
+      result = fakeAPI(length, undefined, searchQuery)
     } else {
-      result = fakeLearnAPI(length)
+      result = fakeAPI(length, data)
     }
 
     result.then(query => this.setState({ query }))
@@ -118,7 +137,7 @@ class CoursesContainer extends React.Component {
     //immitation of search query
     if (this.state.searchQuery) {
       const searchStr = this.state.searchQuery
-      fakeLearnAPISearch(searchStr).then(query => this.setState({ query }))
+      fakeAPISearch(searchStr).then(query => this.setState({ query }))
     }
   }
 
@@ -140,6 +159,11 @@ class CoursesContainer extends React.Component {
           search={this.handleSearch}
           input={this.handleInput}
         />
+        <JobsFilter
+          titles={this.state.filter}
+          onChange={this.handleChange}
+          onSubmit={this.submitQuery}
+        />
         {loading ? (
           <LoadingContent area="lst">Loading...</LoadingContent>
         ) : (
@@ -148,7 +172,6 @@ class CoursesContainer extends React.Component {
               <JobsList
                 key={index}
                 area="lst"
-                type="list"
                 title={elem.title}
                 text={elem}
               />
@@ -165,12 +188,9 @@ class CoursesContainer extends React.Component {
   }
 }
 
-CoursesContainer.propTypes = {
-  menuFilter: PropTypes.array.isRequired
-}
+export default withTheme(JobsContainer)
 
-export default withTheme(CoursesContainer)
-
-CoursesContainer.propTypes = {
+JobsContainer.propTypes = {
+  menuFilter: PropTypes.array.isRequired,
   theme: PropTypes.oneOfType([PropTypes.func, PropTypes.object])
 }
