+++ src/components/ProjectCard.js
@@ -8,7 +8,7 @@ import {
   StyledUl,
   StyledLi,
   media
-} from '../../theme/globalStyle'
+} from '../theme/globalStyle'
 
 import faker from 'faker'
 
@@ -106,9 +106,11 @@ const CardLi = StyledLi.extend`
 
 const ProjectLi = props => (
   <CardUl lang="en">
-    <CardLi>{props.data[0] || faker.date.future().toTimeString()}</CardLi>
-    <CardLi>{`${props.data[1] || faker.address.city()}, ${props.data[2] ||
-      faker.address.country()}`}</CardLi>
+    <CardLi>
+      {props.data[0] || faker.date.future().toTimeString()}
+    </CardLi>
+    <CardLi>{`${props.data[1] || faker.address.city()}, ${props
+      .data[2] || faker.address.country()}`}</CardLi>
     <CardLi>{props.text}</CardLi>
   </CardUl>
 )
