+++ src/components/StoriesContainer.js
@@ -3,13 +3,13 @@ import styled from 'styled-components'
 import { withTheme } from 'styled-components'
 import PropTypes from 'prop-types'
 
-import { media } from '../../theme/globalStyle'
+import { media } from '../theme/globalStyle'
 import FilterAndSearch from './FilterAndSearch'
 import CardContainer from './CardContainer'
-import ProjectCard from './ProjectCard'
-import Pagination from './Pagination'
+import StoryCard from './StoryCard'
 import { LoadingContent } from './Titles'
-import { fakeEventsAPI, fakeEventsAPISearch } from '../utils/api'
+import Pagination from './Pagination'
+import { fakeStoriesAPI, fakeStoriesAPISearch } from '../utils/api'
 
 const Wrapper = styled.div`
   display: grid;
@@ -39,7 +39,7 @@ const Container = styled.div`
   align-content: center;
 `
 
-class EventsContainer extends React.Component {
+class StoriesContainer extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
@@ -49,14 +49,11 @@ class EventsContainer extends React.Component {
         {
           title: '',
           image: '',
-          date: '',
-          city: '',
-          country: '',
           description: ''
         }
       ]
     }
-    this.CARDS = { cols: 5, items: 6 }
+    this.CARDS = { cols: 2, items: 6 }
 
     this.changePage = this.changePage.bind(this)
     this.handleDates = this.handleDates.bind(this)
@@ -66,7 +63,7 @@ class EventsContainer extends React.Component {
 
   componentDidMount() {
     const loading = false
-    fakeEventsAPI(this.CARDS.items).then(query =>
+    fakeStoriesAPI(this.CARDS.items).then(query =>
       this.setState({ loading, query })
     )
   }
@@ -78,9 +75,9 @@ class EventsContainer extends React.Component {
       let result
       if (this.state.searchQuery) {
         const searchQuery = this.state.searchQuery
-        result = fakeEventsAPI(length, undefined, undefined, searchQuery)
+        result = fakeStoriesAPI(length, searchQuery)
       } else {
-        result = fakeEventsAPI(length)
+        result = fakeStoriesAPI(length)
       }
       result.then(query => this.setState({ query }))
     }
@@ -99,19 +96,18 @@ class EventsContainer extends React.Component {
 
   handleDates(str) {
     // immitation of new query
+    const random = Math.floor(Math.random() * (this.CARDS.items + 1))
     const length =
-        str === 'all'
-          ? this.CARDS.items
-          : Math.floor(Math.random() * (this.CARDS.items + 1)),
-      city = str === 'my area' ? 'My City' : '',
-      country = str === 'my area' ? 'My Country' : ''
+      str === 'all'
+        ? this.CARDS.items
+        : random % 2 === 0 ? random : random + 1
     let result
 
     if (this.state.searchQuery) {
       const searchQuery = this.state.searchQuery
-      result = fakeEventsAPI(length, city, country, searchQuery)
+      result = fakeStoriesAPI(length, searchQuery)
     } else {
-      result = fakeEventsAPI(length, city, country)
+      result = fakeStoriesAPI(length)
     }
 
     result.then(query => this.setState({ query }))
@@ -121,7 +117,7 @@ class EventsContainer extends React.Component {
     //immitation of search query
     if (this.state.searchQuery) {
       const searchStr = this.state.searchQuery
-      fakeEventsAPISearch(searchStr, this.CARDS.items).then(query =>
+      fakeStoriesAPISearch(searchStr, this.CARDS.items).then(query =>
         this.setState({ query })
       )
     }
@@ -149,15 +145,13 @@ class EventsContainer extends React.Component {
           <LoadingContent area="cont">Loading...</LoadingContent>
         ) : (
           <Container>
-            <CardContainer cols={this.CARDS.cols} cards={this.CARDS.items}>
+            <CardContainer cols={this.CARDS.cols} story={true}>
               {arr.map((elem, index) => (
-                <ProjectCard
+                <StoryCard
                   key={index}
-                  type="list"
                   title={elem.title}
                   text={elem.description}
                   img={elem.image}
-                  list={[elem.date, elem.city, elem.country]}
                 />
               ))}
             </CardContainer>
@@ -172,9 +166,12 @@ class EventsContainer extends React.Component {
     )
   }
 }
-export default withTheme(EventsContainer)
+export default withTheme(StoriesContainer)
 
-EventsContainer.propTypes = {
+StoriesContainer.propTypes = {
   menuFilter: PropTypes.array.isRequired,
-  theme: PropTypes.PropTypes.oneOfType([PropTypes.func, PropTypes.object])
+  theme: PropTypes.PropTypes.oneOfType([
+    PropTypes.func,
+    PropTypes.object
+  ])
 }
