+++ src/pages/events.js
@@ -1,11 +1,13 @@
 import React from 'react'
-import Hero from '../layouts/components/Hero'
-import EventsContainer from '../layouts/components/EventsContainer'
+import Hero from '../components/Hero'
+import EventsContainer from '../components/EventsContainer'
 
 const EventsPage = ({ data }) => (
   <div>
     <Hero type="events" />
-    <EventsContainer menuFilter={data.site.siteMetadata.filterOptions.events} />
+    <EventsContainer
+      menuFilter={data.site.siteMetadata.filterOptions.events}
+    />
   </div>
 )
