+++ src/pages/learn.js
@@ -2,17 +2,17 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import faker from 'faker'
 import styled from 'styled-components'
-import Hero from '../layouts/components/Hero'
+import Hero from '../components/Hero'
 import { withTheme } from 'styled-components'
 
 import { StyledP, media } from '../theme/globalStyle'
-import CardContainer from '../layouts/components/CardContainer'
-import CoursesContainer from '../layouts/components/CoursesContainer'
-import { addCards } from '../layouts/utils/helpers'
-import Divider from '../layouts/components/Divider'
+import CardContainer from '../components/CardContainer'
+import CoursesContainer from '../components/CoursesContainer'
+import { addCards } from '../utils/helpers'
+import Divider from '../components/Divider'
 import { ICONS } from '../theme/constants'
-import Icon from '../layouts/components/Icon'
-import { SectionTitle } from '../layouts/components/Titles'
+import Icon from '../components/Icon'
+import { SectionTitle } from '../components/Titles'
 
 const IconWrapper = styled.div`
   min-width: 7em;
